requirements.txt
asgiref==3.8.1
asttokens==2.4.1
# decorator==5.1.1
numba==0.60.0
numpy==2.0.2
Django==5.1.2
django-environ==0.11.2
executing==2.1.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
transformers==4.38.2
torch==2.2.0
sentencepiece==0.2.0
protobuf==4.25.2
pytorch-transformers==1.2.0
sacremoses==0.1.1
Pillow==10.2.0
librosa==0.10.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
setuptools
numpy<2.0
openai==1.53.0
openai-whisper==20240930
moviepy==1.0.3
opencv-python==4.10.0.84

